fix: 직무별 랭킹과 전체 랭킹을 정상적으로 가져오도록 수정

+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/spec/dto/response/RankingResponse.java
@@ -24,7 +24,7 @@ public class RankingResponse {
         private Long specId;
         private String jobField;
         private double totalAnalyzeScore;
-        private int rankByJobField;
+        private Long rankByJobField;
         private int totalUsersCountByJobField;
         private long rank;
         private boolean isBookmarked;
