+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/spec/dto/response/SearchResponse.java
@@ -25,7 +25,7 @@ public class SearchResponse {
         private Long specId;
         private String jobField;
         private double totalAnalyzeScore;
-        private int rankByJobField;
+        private Long rankByJobField;
         private int totalUsersCountByJobField;
         private Long rank;
         private boolean isBookmarked;
